+++ tests/test-timmy.php
@@ -75,4 +75,83 @@ class TestTimmy extends TimmyUnitTestCase {
 
 		$this->assertEquals( $expected, $result );
 	}
+
+	/**
+	 * Test with when an attachment is assigned to a post that was already delete.
+	 *
+	 * @link https://github.com/mindkomm/timmy/issues/39
+	 */
+	function test_attached_image_with_missing_post() {
+		$post_id = $this->factory()->post->create();
+		wp_delete_post( $post_id, true );
+
+		$attachment = $this->create_image( [
+			'post_parent' => $post_id,
+		] );
+
+		$expected = sprintf(
+			' srcset="%1$s/test-560x0-c-default.jpg 560w, %1$s/test-1400x0-c-default.jpg 1400w" src="data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7" sizes="100vw" alt=""',
+			$this->get_upload_url()
+		);
+
+		$result = get_timber_image_responsive( $attachment, 'large' );
+
+		$this->assertEquals( $expected, $result );
+	}
+
+	public function test_timmy_ignores_pdf() {
+		$attachment = $this->create_image( [ 'file' => 'example.pdf' ] );
+		$result     = image_downsize( $attachment->ID );
+
+		$image = false;
+		$this->assertEquals( $image, $result );
+	}
+
+	public function test_timmy_ignores_sgv() {
+		$filter = function( $ignore) {
+			$this->assertEquals( true, $ignore );
+
+			return $ignore;
+		};
+
+		add_filter( 'timmy/resize/ignore', $filter, 99 );
+
+		$attachment = $this->create_image( [ 'file' => 'sveegee.svg' ] );
+		image_downsize( $attachment->ID );
+
+		remove_filter( 'timmy/resize/ignore', $filter, 99 );
+	}
+
+	public function test_timmy_ignores_video() {
+		$attachment = $this->create_image( [ 'file' => 'video.mp4' ] );
+		$result     = image_downsize( $attachment->ID );
+
+		$image = false;
+
+		$this->assertEquals( $image, $result );
+	}
+
+	public function test_timmy_ignores_svg_when_generating_metadata() {
+		$attachment = $this->create_image( [ 'file' => 'sveegee.svg' ] );
+		$file_src   = get_attached_file( $attachment->ID );
+		$meta       = wp_generate_attachment_metadata( $attachment->ID, $file_src );
+
+		$this->assertArrayNotHasKey( 'sizes', $meta );
+	}
+
+	public function test_timmy_ignores_gif_when_generating_metadata() {
+		$attachment = $this->create_image( [ 'file' => 'logo-small.gif' ] );
+		$file_src   = get_attached_file( $attachment->ID );
+		$meta       = wp_generate_attachment_metadata( $attachment->ID, $file_src );
+
+		$this->assertEmpty( $meta['sizes'] );
+	}
+
+	public function test_timmy_ignores_video_when_generating_metadata() {
+		$attachment = $this->create_image( [ 'file' => 'video.mp4' ] );
+		$file_src   = get_attached_file( $attachment->ID );
+		$meta       = wp_generate_attachment_metadata( $attachment->ID, $file_src );
+
+		$this->assertArrayNotHasKey( 'sizes', $meta );
+	}
 }