+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## 0.14.9
+
+- Added `Timmy\Timmy::VERSION` constant. You can use this to check Timmy’s version using `version_compare()`:
+    ```php
+    if ( version_compare( Timmy\Timmy::VERSION, '1.0.0', '<' ) ) {
+        // Do something for Timmy versions < 1.0.0.
+    }
+    ```
+- Fixed a bug when Timmy tries to resize files that it shouldn’t, like video files.
+- Updated testing suite.
+- Updated `composer/installers` Composer dependency to allow versions `^1.0` and `^2.0`.
+
+## 0.14.8 - 2021-07-28
+
+- Fixed more issues with regex for Responsive Content Images.
+
+## 0.14.7 - 2021-07-01
+
+- Fixed regex for Responsive Content Images.
+- Fixed a bug for attachments that were deleted from the media library.
+
+## 0.14.6 - 2021-03-29
+
+- Added `content_filters` argument for Responsive Content Images constructor.
+- Fixed performance when a lot of image data is requested.
+
 ## 0.14.5 - 2020-09-11
 
 - Added Block Editor support for the Responsive Content Images functionality.